requirements.txt
argon2-cffi==20.1.0
autopep8==1.5.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pika==1.2.0
pycodestyle==2.7.0
pycparser==2.20
pymongo==3.11.4
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
twilio==6.53.0
pika==1.2.0